Many changes for preprocessor

+++ src/main/java/com/ksilisk/sapr/dto/BarDTO.java
@@ -15,8 +15,7 @@ public class BarDTO {
         this.specInd = specInd;
     }
 
-
-    public double getSpecInd() {
+    public int getSpecInd() {
         return specInd;
     }
